Rebasing to 17.11.12.  We are also removing the restart in the post.
The reason for the restart in post removal is due to the fact that this can lead to errant upgrades.  Namely if I run yum update it will automatically restart the service even though I want to do that together.  This is especially true for slurmdbd upgrades.  So we took that part out.